@@ -63,19 +63,19 @@
     "@jscutlery/semver": "^2.30.1",
     "@nrwl/devkit": "^16.0.1",
     "@types/eslint": "^8.4.8",
-    "@typescript-eslint/eslint-plugin": "^5.47.0",
-    "@typescript-eslint/parser": "^5.47.0",
-    "@typescript-eslint/typescript-estree": "^5.47.0",
+    "@typescript-eslint/eslint-plugin": "^5.62.0",
+    "@typescript-eslint/parser": "^5.62.0",
+    "@typescript-eslint/typescript-estree": "^5.62.0",
     "eslint": "^8.28.0",
     "eslint-config-react-app": "^7.0.1",
-    "eslint-import-resolver-typescript": "^3.5.2",
-    "eslint-plugin-import": "^2.26.0",
+    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-plugin-import": "^2.28.1",
     "eslint-plugin-prettier": "^4.2.1",
     "ngx-deploy-npm": "^5.2.0",
     "nx": "16.0.1",
-    "prettier": "^2.8.4",
-    "typescript": "^4.9.5",
-    "vercel": "^32.1.0",
-    "vite": "^4.4.9"
+    "prettier": "^2.8.8",
+    "typescript": "^5.2.2",
+    "vercel": "^32.4.1",
+    "vite": "^4.4.11"
   }
 }
