@@ -1,13 +1,14 @@
 {
   "name": "@stream-io/i18n",
   "packageManager": "yarn@3.4.1",
-  "main": "dist/index.js",
+  "main": "dist/index.cjs.js",
+  "module": "dist/index.es.js",
   "types": "dist/index.d.ts",
   "license": "See license in LICENSE",
   "scripts": {
     "clean": "rimraf dist",
-    "start": "tsc --project tsconfig.json --watch",
-    "build": "yarn clean && tsc --project tsconfig.json"
+    "start": "rollup -c -w",
+    "build": "yarn clean && rollup -c"
   },
   "files": [
     "dist",
@@ -22,8 +23,10 @@
     "i18next": "^22.4.15"
   },
   "devDependencies": {
-    "rimraf": "^5.0.1",
-    "typescript": "^4.9.5"
+    "@rollup/plugin-typescript": "^11.1.5",
+    "rimraf": "^5.0.5",
+    "rollup": "^3.29.4",
+    "typescript": "^5.2.2"
   },
   "version": "0.1.2"
 }
