@@ -123,7 +123,7 @@
     "@tsconfig/node14": "14.1.0",
     "@types/jest": "^29.5.1",
     "@types/lodash.merge": "^4",
-    "@types/react": "18.0.26",
+    "@types/react": "18.2.28",
     "@types/react-native": "^0.70.4",
     "@types/react-native-incall-manager": "^3.2.1",
     "@types/react-test-renderer": "^18",
@@ -142,10 +142,10 @@
     "react-native-svg": "^13.6.0",
     "react-native-voip-push-notification": "3.3.1",
     "react-test-renderer": "^18.2.0",
-    "rimraf": "^5.0.1",
+    "rimraf": "^5.0.5",
     "ts-jest": "^29.1.0",
     "ts-node": "^10.9.1",
-    "typescript": "^4.9.5"
+    "typescript": "^5.2.2"
   },
   "react-native-builder-bob": {
     "source": "src",
