@@ -13,14 +13,14 @@
     "@stream-io/video-react-sdk": "workspace:^",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.11.2"
+    "react-router-dom": "^6.16.0"
   },
   "devDependencies": {
-    "@types/react": "^18.0.28",
-    "@types/react-dom": "^18.0.11",
-    "@vitejs/plugin-react": "^4.0.0",
-    "typescript": "^4.9.5",
-    "vercel": "^32.1.0",
-    "vite": "^4.4.9"
+    "@types/react": "^18.2.28",
+    "@types/react-dom": "^18.2.13",
+    "@vitejs/plugin-react": "^4.1.0",
+    "typescript": "^5.2.2",
+    "vercel": "^32.4.1",
+    "vite": "^4.4.11"
   }
 }
