@@ -7,7 +7,7 @@ import { Call } from '../../Call';
 const apiKey = process.env.STREAM_API_KEY!;
 const secret = process.env.STREAM_SECRET!;
 
-describe.skip('call members API', () => {
+describe('call members API', () => {
   let client: StreamVideoServerClient;
   const callId = `call${generateUUIDv4()}`;
   let call: Call;
